clean up sys and socket crates
