+++ crates/pheasant_socket/examples/server.rs
@@ -1,9 +1,8 @@
 use pheasant_socket::{
-    Error,
+    AddressFamily, Error, ProtocolNumber, SocketType,
     address::inet::SockAddrIn,
     socket::{SetSockOpts, Socket},
 };
-use pheasant_sys::{AddressFamily, ProtocolNumber, SocketType};
 
 fn main() -> Result<(), Error> {
     let socket = Socket::new(
