+++ crates/pheasant_socket/src/address.rs
@@ -1,3 +1,5 @@
+pub use pheasant_sys::socket::SockAddr;
+
 pub mod inet;
 // pub mod inet6;
 pub mod unix;
@@ -10,26 +12,6 @@ pub use unix::SockAddrUn;
 /// the passed address boundaries of the sockaddr_* struct
 /// WARN (move this somewhere else) the address port mst be passed in in big endianne notation
 /// else the socket would bind to the be repr of your le number
-
-#[repr(C)]
-#[derive(Debug)]
-pub struct SockAddr {
-    // use AddressFamily::ChoiceFamily.as_int() to populate this field
-    // as it refers to a c_uint repr of the address family value
-    family: u16,
-    data: [u8; 110],
-}
-
-impl SockAddr {
-    pub const SIZE: u32 = core::mem::size_of::<Self>() as u32;
-}
-
-// handles both ipv4 and 6
-// #[repr(C)]
-// pub struct SockAddrStorage {
-//     // address family int repr
-//     ss_family: u16,
-// }
 
 #[cfg(test)]
 mod sizes {
